pyproject.toml
[project]
name = 'aircraft-maintenance-rag'
version = '0.2.0'
description = 'A multi-agent RAG system for aircraft maintenance documentation using LangChain, LangGraph, LangSmith, Ollama.'
readme = 'README.md'
requires-python = '>=3.11'
dependencies = [
    'chromadb>=1.3.5',
    'langchain>=1.1.0',
    "langchain-chroma>=1.0.0",
    'langchain-community>=0.4.1',
    "langchain-huggingface>=1.1.0",
    "langchain-openai>=1.1.0",
    "openai>=2.9.0",
    "pymupdf>=1.26.6",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=5.1.2",
    'streamlit>=1.51.0',
]
